Fix discord login issues
I am not sure 100% what is going on on production as I can't reproduce
in local env. My theory is that vercel function does not play nice with
async methods that are not awaited (?). This PR adds await to create
Event.create in order to try to fix the login issue.